Cast is not needed.

+++ src/df1b2-separable/df1b2lp8.cpp
@@ -169,7 +169,7 @@ void laplace_approximation_calculator::generate_antithetical_rvs()
   int ii;
   for (x=mid-spread;x<=mid+spread;x+=delta)
   {
-    tsum+=exp((n-1)*log(x)-0.5*x*x-tmax)/ssum*((double)samplesize);
+    tsum+=exp((n-1)*log(x)-0.5*x*x-tmax)/ssum*samplesize;
     int ns=int(tsum);
     for (ii=1;ii<=ns;ii++)
     {
